requirements.txt
django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
httpie # python -m pip install httpie     better use postman

